Fix errors of messages from hidden accounts
Now bot is using Sqlite3 with peewee ORM. Docker is not tested (and is
not configured) so its notice was removed.

Every message's from_user and blocking users are saved in db.sqlite3 file.

Sqlite was chosen because I was working with peewee and it showed itself as a
good ORM. It could be better to use something simpler as LMDB but it's new for
me.

And I have to go to school right now so I haven't had much time to release and
compare solutions (which weren't).